+++ notary-server/fabric/business/interface.go
