+++ notary-server/fabric/client/client_test.go
@@ -2,7 +2,7 @@ package client
 
 import (
 	"fmt"
-	"github.com/ehousecy/notary-samples/fabric/sdkutil"
+	"github.com/ehousecy/notary-samples/notary-server/fabric/sdkutil"
 	"github.com/golang/protobuf/proto"
 	pb "github.com/hyperledger/fabric-protos-go/peer"
 	"github.com/hyperledger/fabric-sdk-go/pkg/client/channel"
